pyproject.toml
[tool.poetry]
name = "ayayaxyz"
version = "0.1.0"
description = ""
authors = ["tretrauit <tretrauit@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {version = "20.0a4", allow-prereleases = true}
gppt = {version = "^2.2.0", optional = true}
python-dotenv = "^0.20.0"
PixivPy3 = "^3.7.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.extras]
login = ["ipdb"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
